Extended ore gen configuration
