+++ src/main/java/divinerpg/utils/log/PostInitLog.java
@@ -1,8 +1,8 @@
 package divinerpg.utils.log;
 
-import divinerpg.Config;
 import divinerpg.DivineRPG;
 import divinerpg.Reference;
+import divinerpg.config.Config;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
